Moving to utils directory: hooks, momentResolver, Profiler, testUtils. velocityTracker
