+++ src/calendar/index.spec.js
@@ -2,7 +2,7 @@ import XDate from 'xdate';
 import React from 'react';
 import {getTextNodes} from 'react-component-driver';
 import {advanceTo, clear as clearDate} from 'jest-date-mock';
-import {getDaysArray, partial} from '../testUtils';
+import {getDaysArray, partial} from '../utils/testUtils';
 import {CalendarDriver} from './driver';
 
 describe('Calendar', () => {
