+++ src/utils/velocityTracker.ts
